refactor(perl-lsp): split lsp_harness.rs into focused support modules
Extract three domain-specific modules from the 1349-line lsp_harness.rs:

- test_workspace.rs: TempWorkspace struct and helpers (48 lines)
- message_framing.rs: TestWriter and SendableServer (68 lines)
- notification_queue.rs: TestContext compatibility wrapper (130 lines)

lsp_harness.rs remains as the facade, re-exporting all public symbols
so that 150+ downstream test files continue to compile without import
changes.
