+++ crates/perl-lsp/tests/support/mod.rs
@@ -4,7 +4,10 @@ pub mod client_caps;
 pub mod env_guard;
 pub mod lsp_client;
 pub mod lsp_harness;
+pub mod message_framing;
+pub mod notification_queue;
 pub mod test_helpers;
+pub mod test_workspace;
 
 // Re-export test helpers for convenience in test files that use `support::*`
 // NOTE: test_helpers module exists but may not be used in all test contexts
@@ -17,4 +20,4 @@ pub use lsp_harness::{handshake_initialize, shutdown_graceful, spawn_lsp};
 
 // Re-export types that tests may need
 #[allow(unused_imports)]
-pub use lsp_harness::{LspHarness, TempWorkspace};
+pub use lsp_harness::{LspHarness, TempWorkspace, TestContext};
